requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
gunicorn==23.0.0
aiosqlite==0.20.0
asyncpg==0.30.0
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.9.2
